+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Fix bug in development model where `@AutoObserve` for `@Nonnull` fields and methods where the observable value is in the process of being disposed.
+
 ### [v0.237](https://github.com/arez/arez/tree/v0.237) (2026-04-27) · [Full Changelog](https://github.com/arez/arez/compare/v0.236...v0.237)
 
 Changes in this release: